+++ Client/BlockProcessing/Game/Features/Entities/Player/PlayerBlockOutline.h
@@ -10,13 +10,13 @@ class PlayerBlockOutline {
 public:
     void init();
     void update(int64_t x, int64_t y, int64_t z);
-    void render(Mat4f& view);
+    void render(Mat4d& view);
     void setProjection(Mat4f& projection);
     UIColor color;
     ~PlayerBlockOutline();
 private:
     Shader* shader;
-    Mat4f model;
+    Mat4d model;
     VertexArrayObject vao;
     VertexBufferObject vbo;
     int vertexCount;
